RT#86041: fix coercion result on failure
attempt to load class in the validation itself, not the coercion, so we can
fail out properly, without returning an undef result when attempting to coerce
a bad string.
It's MUCH easier to simply test with the Moose::Meta::TypeConstraint API,
rather than attempting to shoehorn into a Moose class.